+++ CHANGELOG.md
@@ -49,6 +49,8 @@ This is a mandatory release for calibnet node operators. It includes the revised
 
 - [#5386](https://github.com/ChainSafe/forest/pull/5386) Add support for the `Filecoin.EthTraceTransaction` RPC method.
 
+- [#5427](https://github.com/ChainSafe/forest/pull/5427) Add support for the `Filecoin.EthTraceFilter` RPC method.
+
 - [#5383](https://github.com/ChainSafe/forest/pull/5383) Add support for `Filecoin.EthGetFilterChanges` RPC method.
 
 ### Fixed